Improve Scheduler loggers. Add two new appenders that store logs in a file.
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/scheduling/trunk@11405 28e8926c-6b08-0410-baaa-805c5e19b8d6

+++ bin/unix/startScheduler_newDB.sh
@@ -1,10 +1,10 @@
 #!/bin/sh
 
+if [ -d SCHEDULER_DB ]
+then
 rm -rf SCHEDULER_DB
 rm derby.log
-
-echo
-echo --- Scheduler----------------------------------------------
+fi
 
 
 workingDir=`pwd`
@@ -14,7 +14,6 @@ CLASSPATH=.
 
 yjp=-agentlib:yjpagent
 opt="-Xms128m -Xmx512m"
-
 
 $JAVACMD $opt org.ow2.proactive.scheduler.examples.SchedulerStarter $@
